Check for CVEs in Oracle InstantClient during build (#124)
Check OSV and NIST NVD to see if Oracle InstantClient
has had any CVEs reported. (As of 2023, none ever have been,
so the check is our best guess as to what name they will be published
under.)

Initially was going to add `security-scanner` to scan the binary
as well, but it looks like it does not support Cgo binaries.

Also updated Go to 1.20.4.